+++ CHANGELOG.md
@@ -15,7 +15,7 @@ as of 2.0.0.
 - Fix issue with i18n fallback language (#360)
 - Convert to WEBP and optimize WEBP, JPG and PNG images (both HTML resources + cover image) (#375)
 - Optimize Docker build topology to reuse layers (#420)
-- Optimize ePub again (#376)
+- Optimize ePub again (#376, #436)
 - Fail scraper when RDF download fails (#414)
 
 ### Changed